Remove BasePipeline and make a module for RayPipeline
